# Arogya old/docs/API_SPEC.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AROGYA Healthcare Platform API",
    "description": "Production-ready REST API for rural healthcare platform",
    "version": "1.0.0",
    "contact": {
      "name": "AROGYA Team",
      "email": "api@arogya.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.arogya.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/doctors": {
      "get": {
        "tags": ["Doctors"],
        "summary": "List doctors",
        "description": "Get list of doctors with optional filtering",
        "parameters": [
          {
            "name": "hospital_id",
            "in": "query",
            "description": "Filter by hospital",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "specialty",
            "in": "query",
            "description": "Filter by specialty",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of doctors"
          }
        }
      }
    },
...
    "/api/v1/reminders": {
      "post": {
        "tags": ["Reminders"],
        "summary": "Create medication reminder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["medication_name", "dosage", "frequency", "reminder_times"],
                "properties": {
                  "medication_name": {
                    "type": "string"
                  },
                  "dosage": {
                    "type": "string",
                    "example": "500mg"
                  },
                  "frequency": {
                    "type": "string",
                    "example": "Once daily"
                  },
                  "reminder_times": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "time",
                      "example": "08:00"
                    }
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reminder created"
          }
        }
      },
      "get": {
        "tags": ["Reminders"],
        "summary": "Get user reminders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of reminders"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token"
      }
    },
...
  },
...
}